src/backend/requirements.txt
fastapi>=0.135.0
uvicorn[standard]>=0.41.0
pydantic-settings>=2.7.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
redis>=5.1.0
hiredis>=2.3.2
celery>=5.4.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
geoalchemy2>=0.14.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
respx>=0.21.0
python-keycloak>=4.0.0
pyyaml>=6.0.1
